fix flickering on height
