+++ src/components/simulator/result/SimResultChartAnimation.tsx
@@ -53,14 +53,14 @@ export const SimResultChartAnimation = ({
     <>
       <D3ChartWrapper
         settings={chartSettings}
-        className="w-full border-t border-emphasis"
+        className="min-h-[350px] w-full border-t border-emphasis"
       >
         {(dms) => (
           <D3ChartSimulatorPrice data={data} bounds={bounds!} dms={dms} />
         )}
       </D3ChartWrapper>
 
-      <div className={`grid ${gridCols}`}>
+      <div className={`grid ${gridCols} min-h-[250px] items-stretch`}>
         <D3ChartWrapper
           settings={chartPerformanceSettings}
           className="w-full border-t border-r border-emphasis"
